Move erbui to its own python module
