+++ build-system/cover.py
@@ -70,8 +70,8 @@ def main ():
 
    erase_coverage_data ()
 
-   collect ('erbb.tests.test_parser')
-   collect ('erbb.tests.test_generators')
+   collect ('erbui.tests.test_parser')
+   collect ('erbui.tests.test_generators')
 
    report ()
