+++ build-system/erbui/generators/detail/panel.py
