+++ build-system/erbui/generators/front_panel/__init__.py
