+++ build-system/erbui/generators/vcvrack/manifest.py
